cw2/requirements.txt
Flask>=3.1,<4
flask-cors>=6,<7
PyJWT>=2.10,<3
Flask-Limiter>=3.8,<4
bcrypt>=4,<5
pyotp>=2.9,<3
